+++ builder/amazon/common/access_config.go
@@ -17,6 +17,7 @@ import (
 	"github.com/aws/aws-sdk-go/service/ec2/ec2iface"
 	awsbase "github.com/hashicorp/aws-sdk-go-base"
 	cleanhttp "github.com/hashicorp/go-cleanhttp"
+	"github.com/hashicorp/packer/builder/amazon/common/awserrors"
 	"github.com/hashicorp/packer/template/interpolate"
 	vaultapi "github.com/hashicorp/vault/api"
 )
@@ -271,7 +272,7 @@ func (c *AccessConfig) Session() (*session.Session, error) {
 
 	cp, err := c.session.Config.Credentials.Get()
 
-	if IsAWSErr(err, "NoCredentialProviders", "") {
+	if awserrors.Matches(err, "NoCredentialProviders", "") {
 		return nil, c.NewNoValidCredentialSourcesError(err)
 	}
