+++ builder/amazon/common/awserrors/utils.go
@@ -1,4 +1,4 @@
-package error
+package awserrors
 
 import (
 	"strings"
